@@ -1,10 +1,11 @@
 apply plugin: 'com.android.library'
 
 android {
-    compileSdk 32
+
     defaultConfig {
-        minSdk 23
-        targetSdk 32
+        minSdk 26
+        targetSdk 33
+        compileSdk 33
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
     sourceSets {
@@ -29,15 +30,17 @@ android {
         sourceCompatibility JavaVersion.VERSION_11
         targetCompatibility JavaVersion.VERSION_11
     }
+    namespace 'com.ebook.basebook'
 }
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     api project(':lib_common')
-    api 'androidx.appcompat:appcompat:1.4.1'
+    api 'androidx.appcompat:appcompat:1.6.1'
     testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.3'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.5'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
     //权限 Permission x
-    api 'com.permissionx.guolindev:permissionx:1.4.0'
+    api 'com.guolindev.permissionx:permissionx:1.7.1'
+    implementation 'com.github.promeg:tinypinyin:2.0.3'
 }
