@@ -1,10 +1,11 @@
 apply plugin: 'com.android.library'
 
 android {
-    compileSdk 32
+
     defaultConfig {
-        minSdkVersion 23
-        targetSdk 32
+        minSdk 26
+        targetSdk 33
+        compileSdk 33
         javaCompileOptions {
             annotationProcessorOptions {
                 arguments = [AROUTER_MODULE_NAME: project.getName()]
@@ -30,6 +31,7 @@ android {
         sourceCompatibility JavaVersion.VERSION_11
         targetCompatibility JavaVersion.VERSION_11
     }
+    namespace 'com.ebook.common'
 }
 
 dependencies {
@@ -39,16 +41,16 @@ dependencies {
     api project(':lib_ebook_api')
     api project(":lib_ebook_db")
     api 'androidx.legacy:legacy-support-v4:1.0.0'
-    api 'androidx.appcompat:appcompat:1.4.1'
+    api 'androidx.appcompat:appcompat:1.6.1'
     api 'androidx.constraintlayout:constraintlayout:2.1.4'
-    api 'com.google.android.material:material:1.6.0'
+    api 'com.google.android.material:material:1.8.0'
     api 'com.alibaba:arouter-api:1.5.2'
     annotationProcessor 'com.alibaba:arouter-compiler:1.5.2'
     //rxjava
     api 'io.reactivex.rxjava2:rxjava:2.2.12'
     api 'io.reactivex.rxjava2:rxandroid:2.1.1'
 
-    api 'org.greenrobot:eventbus:3.2.0'
+    api 'org.greenrobot:eventbus:3.3.1'
 
     api('com.jakewharton.rxbinding2:rxbinding:2.2.0') {
         exclude group: 'com.android.support'
@@ -60,7 +62,7 @@ dependencies {
     annotationProcessor 'com.github.bumptech.glide:compiler:4.13.1'
 
     api 'androidx.lifecycle:lifecycle-extensions:2.2.0'
-    annotationProcessor 'androidx.lifecycle:lifecycle-common-java8:2.4.1'
+    annotationProcessor 'androidx.lifecycle:lifecycle-common-java8:2.6.1'
 
     api 'com.google.dagger:dagger:2.37'
     annotationProcessor 'com.google.dagger:dagger-compiler:2.37'
@@ -70,8 +72,8 @@ dependencies {
     api 'com.github.lovetuzitong:MultiImageSelector:1.2'
 
     testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.3'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.5'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
     // 安卓工具类 https://github.com/Blankj/AndroidUtilCode
     api 'com.blankj:utilcodex:1.31.0'
 
