@@ -1,10 +1,11 @@
 apply plugin: 'com.android.library'
 
 android {
-    compileSdk 32
+
     defaultConfig {
-        minSdk 23
-        targetSdk 32
+        minSdk 26
+        targetSdk 33
+        compileSdk 33
     }
     sourceSets {
         main {
@@ -29,11 +30,12 @@ android {
         sourceCompatibility JavaVersion.VERSION_11
         targetCompatibility JavaVersion.VERSION_11
     }
+    namespace 'com.ebook.api'
 }
 
 dependencies {
     api fileTree(include: ['*.jar'], dir: 'libs')
-    api 'androidx.appcompat:appcompat:1.4.1'
+    api 'androidx.appcompat:appcompat:1.6.1'
     api 'org.jetbrains:annotations:23.0.0'
 
     //network
@@ -44,6 +46,7 @@ dependencies {
 
     //json解析
     api 'com.google.code.gson:gson:2.8.9'
+    api 'com.alibaba.fastjson2:fastjson2:2.0.27'
     //rx管理View的生命周期
     api('com.trello.rxlifecycle3:rxlifecycle:3.1.0') {
         exclude group: 'com.android.support'
@@ -56,8 +59,8 @@ dependencies {
     }
 
     testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.3'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.5'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
 
     api 'com.squareup.retrofit2:converter-scalars:2.9.0'
 }
