@@ -1,11 +1,11 @@
 apply plugin: 'com.android.library'
 apply plugin: 'org.greenrobot.greendao'
 android {
-    compileSdk 32
-    defaultConfig {
-        minSdk 23
-        targetSdk 32
 
+    defaultConfig {
+        minSdk 26
+        targetSdk 33
+        compileSdk 33
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         consumerProguardFiles 'proguard-rules.pro'
     }
@@ -31,18 +31,19 @@ android {
         sourceCompatibility JavaVersion.VERSION_11
         targetCompatibility JavaVersion.VERSION_11
     }
+    namespace 'com.ebook.db'
 
 }
 
 dependencies {
     api fileTree(dir: 'libs', include: ['*.jar'])
-    api 'androidx.appcompat:appcompat:1.4.1'
+    api 'androidx.appcompat:appcompat:1.6.1'
 
     api 'org.greenrobot:greendao:3.3.0'
     api 'org.greenrobot:greendao-generator:3.3.0'
     api 'org.greenrobot:greendao-gradle-plugin:3.3.0'
 
     testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.3'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.5'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
 }
