@@ -1,10 +1,11 @@
 apply plugin: 'com.android.library'
 
 android {
-    compileSdk 32
+
     defaultConfig {
         minSdkVersion 23
-        targetSdk 32
+        targetSdk 33
+        compileSdk 33
     }
     sourceSets {
         main {
@@ -28,14 +29,15 @@ android {
         sourceCompatibility JavaVersion.VERSION_11
         targetCompatibility JavaVersion.VERSION_11
     }
+    namespace 'com.github.lib'
 }
 
 dependencies {
     api fileTree(include: ['*.jar'], dir: 'libs')
-    api 'androidx.appcompat:appcompat:1.4.1'
-    api 'androidx.recyclerview:recyclerview:1.2.1'
+    api 'androidx.appcompat:appcompat:1.6.1'
+    api 'androidx.recyclerview:recyclerview:1.3.0'
 
     testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.3'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.5'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
 }
