@@ -4,10 +4,11 @@ if (isModule.toBoolean()) {
     apply plugin: 'com.android.library'
 }
 android {
-    compileSdk 32
+
     defaultConfig {
-        minSdk 23
-        targetSdk 32
+        minSdk 26
+        targetSdk 33
+        compileSdk 33
         versionCode 3
         versionName '1.1.4'
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
