@@ -7,11 +7,11 @@ package com.ebook.basebook.constant;
  * @date 2022/8/28
  */
 public interface Url {
-    String xh = "https://www.shuangliusc.com/xuanhuanxiaoshuo/";
-    String xz = "https://www.shuangliusc.com/xiuzhenxiaoshuo/";
-    String ds = "https://www.shuangliusc.com/dushixiaoshuo/";
-    String ls = "https://www.shuangliusc.com/lishixiaoshuo/";
-    String wy = "https://www.shuangliusc.com/wangyouxiaoshuo/";
-    String kh = "https://www.shuangliusc.com/kehuanxiaoshuo/";
-    String qt = "https://www.shuangliusc.com/qitaxiaoshuo/";
+    String xh_qh = "http://www.bbzayy.com/sort/1/1.html";
+    String wx_xx = "http://www.bbzayy.com/sort/2/1.html";
+    String ds_yq = "http://www.bbzayy.com/sort/3/1.html";
+    String ls_js = "http://www.bbzayy.com/sort/4/1.html";
+    String kh_ly = "http://www.bbzayy.com/sort/5/1.html";
+    String wy_jj = "http://www.bbzayy.com/sort/6/1.html";
+    String np = "http://www.bbzayy.com/sort/7/1.html";
 }
