@@ -27,7 +27,7 @@ public class WebBookModelImpl implements WebBookModel {
             synchronized (WebBookModelImpl.class) {
                 if (bookModel == null) {
                     //更换书源只需要修改这一行
-                    bookModel = new WebBookModelImpl(BiQuGeBookModelImpl.getInstance());
+                    bookModel = new WebBookModelImpl(ZeroBookModelImpl.getInstance());
                 }
             }
         }
