@@ -1,8 +1,13 @@
 package com.ebook.basebook.mvp.model.impl;
 
+import static com.github.promeg.pinyinhelper.Pinyin.toPinyin;
+
 import android.util.Log;
 
-import com.ebook.api.service.BeQuGeService;
+import com.alibaba.fastjson2.JSONArray;
+import com.alibaba.fastjson2.JSONObject;
+import com.ebook.api.entity.BookEntity;
+import com.ebook.api.service.ZeroBookService;
 import com.ebook.basebook.base.impl.MBaseModelImpl;
 import com.ebook.basebook.base.manager.ErrorAnalyContentManager;
 import com.ebook.basebook.cache.ACache;
@@ -37,17 +42,17 @@ import io.reactivex.schedulers.Schedulers;
 
 /**
  * @author xrn1997
- * @date 2021/6/19
+ * @date 2023/3/27
  */
-public class BiQuGeBookModelImpl extends MBaseModelImpl implements StationBookModel {
-    private volatile static BiQuGeBookModelImpl bookModel;
-    private final String TAG = "shuangliusc.com";
+public class ZeroBookModelImpl extends MBaseModelImpl implements StationBookModel {
+    private volatile static ZeroBookModelImpl bookModel;
+    private final String TAG = "bbzayy.com";
 
-    public static BiQuGeBookModelImpl getInstance() {
+    public static ZeroBookModelImpl getInstance() {
         if (bookModel == null) {
-            synchronized (BiQuGeBookModelImpl.class) {
+            synchronized (ZeroBookModelImpl.class) {
                 if (bookModel == null) {
-                    bookModel = new BiQuGeBookModelImpl();
+                    bookModel = new ZeroBookModelImpl();
                 }
             }
         }
@@ -59,35 +64,35 @@ public class BiQuGeBookModelImpl extends MBaseModelImpl implements StationBookMo
     public Observable<List<SearchBook>> getKindBook(String url, int page) {
         int type = -1;
         switch (url) {
-            case Url.xh:
+            case Url.xh_qh:
                 type = 1;
                 break;
-            case Url.xz:
+            case Url.wx_xx:
                 type = 2;
                 break;
-            case Url.ds:
+            case Url.ds_yq:
                 type = 3;
                 break;
-            case Url.ls:
+            case Url.ls_js:
                 type = 4;
                 break;
-            case Url.wy:
+            case Url.kh_ly:
                 type = 5;
                 break;
-            case Url.kh:
+            case Url.wy_jj:
                 type = 6;
                 break;
-            case Url.qt:
-                type = 8;
+            case Url.np:
+                type = 7;
                 break;
         }
         if (type == -1) {
             Log.e(TAG, "getKindBook: 网址错误");
             return null;
         }
-        return getRetrofitObject(BeQuGeService.URL)
-                .create(BeQuGeService.class)
-                .getKindBooks("/list/" + type + "_" + page + ".html")
+        return getRetrofitObject(ZeroBookService.URL)
+                .create(ZeroBookService.class)
+                .getKindBooks("/sort/" + type + "/" + page + ".html")
                 .flatMap((Function<String, ObservableSource<List<SearchBook>>>) this::analyzeKindBook);
     }
 
@@ -99,20 +104,13 @@ public class BiQuGeBookModelImpl extends MBaseModelImpl implements StationBookMo
             List<SearchBook> books = new ArrayList<>();
             for (int i = 0; i < kindBookEs.size(); i++) {
                 SearchBook item = new SearchBook();
-                item.setTag(BeQuGeService.URL);
+                item.setTag(ZeroBookService.URL);
                 item.setAuthor(kindBookEs.get(i).getElementsByClass("s4").text());
                 item.setLastChapter(kindBookEs.get(i).getElementsByTag("a").get(1).text());
                 item.setOrigin(TAG);
                 item.setName(kindBookEs.get(i).getElementsByTag("a").get(0).text());
                 item.setNoteUrl(kindBookEs.get(i).getElementsByTag("a").get(0).attr("href"));
-                String[] temp = item.getNoteUrl().split("/");
-                char c;
-                if (temp[temp.length - 1].length() == 4) {
-                    c = temp[temp.length - 1].charAt(0);
-                } else {
-                    c = '0';
-                }
-                item.setCoverUrl(BeQuGeService.COVER_URL + "/" + c + "/" + temp[temp.length - 1] + "/" + temp[temp.length - 1] + "s.jpg");
+                item.setCoverUrl(ZeroBookService.COVER_URL + "/" + toPinyin(item.getName(), "").toLowerCase() + ".jpg");
                 books.add(item);
             }
             e.onNext(books);
@@ -123,7 +121,7 @@ public class BiQuGeBookModelImpl extends MBaseModelImpl implements StationBookMo
     /////////////////////////////////////////////////////////////////////////////////////////////////////////////
     @Override
     public Observable<Library> getLibraryData(ACache aCache) {
-        return getRetrofitObject(BeQuGeService.URL).create(BeQuGeService.class).getLibraryData("").flatMap((Function<String, ObservableSource<Library>>) s -> {
+        return getRetrofitObject(ZeroBookService.URL).create(ZeroBookService.class).getLibraryData("").flatMap((Function<String, ObservableSource<Library>>) s -> {
             if (s.length() > 0 && aCache != null) {
                 aCache.put("cache_library", s);
             }
@@ -145,7 +143,7 @@ public class BiQuGeBookModelImpl extends MBaseModelImpl implements StationBookMo
             List<LibraryNewBook> libraryNewBooks = new ArrayList<>();
             for (int i = 0; i < newBookEs.size(); i++) {
                 Element itemE = newBookEs.get(i).getElementsByTag("a").get(0);
-                LibraryNewBook item = new LibraryNewBook(itemE.text(), itemE.attr("href"), BeQuGeService.URL, TAG);
+                LibraryNewBook item = new LibraryNewBook(itemE.text(), itemE.attr("href"), ZeroBookService.URL, TAG);
                 libraryNewBooks.add(item);
             }
             result.setLibraryNewBooks(libraryNewBooks);
@@ -157,11 +155,11 @@ public class BiQuGeBookModelImpl extends MBaseModelImpl implements StationBookMo
             for (int i = 0; i < kindContentEs.size(); i++) {
                 LibraryKindBookList kindItem = new LibraryKindBookList();
                 kindItem.setKindName(kindContentEs.get(i).getElementsByTag("h2").get(0).text());
-                kindItem.setKindUrl(BeQuGeService.URL + kindEs.get(0).getElementsByTag("a").get(i + 2).attr("href"));
+                kindItem.setKindUrl(ZeroBookService.URL + kindEs.get(0).getElementsByTag("a").get(i + 2).attr("href"));
                 List<SearchBook> books = new ArrayList<>();
                 Element firstBookE = kindContentEs.get(i).getElementsByClass("top").get(0);
                 SearchBook firstBook = new SearchBook();
-                firstBook.setTag(BeQuGeService.URL);
+                firstBook.setTag(ZeroBookService.URL);
                 firstBook.setOrigin(TAG);
                 firstBook.setName(firstBookE.getElementsByTag("a").get(1).text());
                 firstBook.setNoteUrl(firstBookE.getElementsByTag("a").get(0).attr("href"));
@@ -171,19 +169,12 @@ public class BiQuGeBookModelImpl extends MBaseModelImpl implements StationBookMo
                 Elements otherBookEs = kindContentEs.get(i).getElementsByTag("li");
                 for (int j = 0; j < otherBookEs.size(); j++) {
                     SearchBook item = new SearchBook();
-                    item.setTag(BeQuGeService.URL);
+                    item.setTag(ZeroBookService.URL);
                     item.setOrigin(TAG);
                     item.setKind(kindItem.getKindName());
                     item.setNoteUrl(otherBookEs.get(j).getElementsByTag("a").get(0).attr("href"));
-                    String[] temp = item.getNoteUrl().split("/");
-                    char c;
-                    if (temp[temp.length - 1].length() == 4) {
-                        c = temp[temp.length - 1].charAt(0);
-                    } else {
-                        c = '0';
-                    }
-                    item.setCoverUrl(BeQuGeService.COVER_URL + "/" + c + "/" + temp[temp.length - 1] + "/" + temp[temp.length - 1] + "s.jpg");
                     item.setName(otherBookEs.get(j).getElementsByTag("a").get(0).text());
+                    item.setCoverUrl(ZeroBookService.COVER_URL + "/" + toPinyin(item.getName(), "").toLowerCase() + ".jpg");
                     books.add(item);
                 }
                 kindItem.setBooks(books);
@@ -200,10 +191,10 @@ public class BiQuGeBookModelImpl extends MBaseModelImpl implements StationBookMo
     @Override
     public Observable<List<SearchBook>> searchBook(String content, int page) {
         try {
-            String str = URLEncoder.encode(content, "GBK");
-            return getRetrofitObject(BeQuGeService.URL)
-                    .create(BeQuGeService.class)
-                    .searchBook(str)
+            String str = URLEncoder.encode(content, "UTF-8");
+            return getRetrofitObject(ZeroBookService.SEARCH_URL)
+                    .create(ZeroBookService.class)
+                    .searchBook(str, page, "bbzayy_com")
                     .flatMap((Function<String, ObservableSource<List<SearchBook>>>) this::analyzeSearchBook);
         } catch (UnsupportedEncodingException e) {
             throw new RuntimeException(e);
@@ -214,27 +205,22 @@ public class BiQuGeBookModelImpl extends MBaseModelImpl implements StationBookMo
     public Observable<List<SearchBook>> analyzeSearchBook(final String s) {
         return Observable.create(e -> {
             try {
-                Document doc = Jsoup.parse(s);
-                Elements booksE = doc.getElementsByAttributeValue("class", "txt-list txt-list-row5").get(0).getElementsByTag("li");
-                //第一个为列表表头，所以如果有书booksE的size必定大于2
-                if (null != booksE && booksE.size() >= 2) {
+                String json = s.substring(1, s.length() - 1);
+                JSONObject jsonObject = JSONObject.parseObject(json);
+                JSONArray jsonArray = jsonObject.getJSONArray("data");
+                List<BookEntity> searchList = jsonArray.toJavaList(BookEntity.class);
+                if (searchList.size() != 0) {
                     List<SearchBook> books = new ArrayList<>();
-                    for (int i = 1; i < booksE.size(); i++) {
+                    for (int i = 0; i < searchList.size(); i++) {
                         SearchBook item = new SearchBook();
-                        item.setTag(BeQuGeService.URL);
-                        item.setAuthor(booksE.get(i).getElementsByClass("s4").get(0).text());
-                        item.setLastChapter(booksE.get(i).getElementsByClass("s3").get(0).getElementsByTag("a").get(0).text());
+                        item.setTag(ZeroBookService.URL);
+                        item.setAuthor(searchList.get(i).getAuthor());
+                        item.setLastChapter(searchList.get(i).getLastChapter());
                         item.setOrigin(TAG);
-                        item.setName(booksE.get(i).getElementsByClass("s2").get(0).getElementsByTag("a").get(0).text());
-                        item.setNoteUrl(booksE.get(i).getElementsByClass("s2").get(0).getElementsByTag("a").get(0).attr("href"));
-                        String[] temp = item.getNoteUrl().split("/");
-                        char c;
-                        if (temp[temp.length - 1].length() == 4) {
-                            c = temp[temp.length - 1].charAt(0);
-                        } else {
-                            c = '0';
-                        }
-                        item.setCoverUrl(BeQuGeService.COVER_URL + "/" + c + "/" + temp[temp.length - 1] + "/" + temp[temp.length - 1] + "s.jpg");
+                        item.setName(searchList.get(i).getName());
+                        item.setNoteUrl(ZeroBookService.URL + "/ldks/" + searchList.get(i).getId() + "/");
+                        item.setCoverUrl(ZeroBookService.COVER_URL + "/" + toPinyin(item.getName(), "").toLowerCase() + ".jpg");
+                        item.setDesc(searchList.get(i).getDesc());
                         books.add(item);
                     }
                     e.onNext(books);
@@ -252,15 +238,15 @@ public class BiQuGeBookModelImpl extends MBaseModelImpl implements StationBookMo
     /////////////////////////////////////////////////////////////////////////////////////////////////////////////
     @Override
     public Observable<BookShelf> getBookInfo(BookShelf bookShelf) {
-        return getRetrofitObject(BeQuGeService.URL)
-                .create(BeQuGeService.class)
-                .getBookInfo(bookShelf.getNoteUrl().replace(BeQuGeService.URL, ""))
+        return getRetrofitObject(ZeroBookService.URL)
+                .create(ZeroBookService.class)
+                .getBookInfo(bookShelf.getNoteUrl().replace(ZeroBookService.URL, ""))
                 .flatMap((Function<String, ObservableSource<BookShelf>>) s -> analyzeBookInfo(s, bookShelf));
     }
 
     private Observable<BookShelf> analyzeBookInfo(String s, BookShelf bookShelf) {
         return Observable.create(e -> {
-            bookShelf.setTag(BeQuGeService.URL);
+            bookShelf.setTag(ZeroBookService.URL);
             bookShelf.setBookInfo(analyzeBookInfo(s, bookShelf.getNoteUrl()));
             e.onNext(bookShelf);
             e.onComplete();
@@ -270,7 +256,7 @@ public class BiQuGeBookModelImpl extends MBaseModelImpl implements StationBookMo
     private BookInfo analyzeBookInfo(String s, String novelUrl) {
         BookInfo bookInfo = new BookInfo();
         bookInfo.setNoteUrl(novelUrl);   //id
-        bookInfo.setTag(BeQuGeService.URL);
+        bookInfo.setTag(ZeroBookService.URL);
         Document doc = Jsoup.parse(s);
         bookInfo.setName(doc.getElementsByClass("info").get(0).getElementsByTag("h1").get(0).text());
         bookInfo.setAuthor(doc.getElementsByClass("info").get(0).getElementsByTag("p").get(0).text().replace("作&nbsp;&nbsp;&nbsp;&nbsp;者：", ""));
@@ -278,14 +264,7 @@ public class BiQuGeBookModelImpl extends MBaseModelImpl implements StationBookMo
         if (bookInfo.getIntroduce().equals("\u3000\u3000")) {
             bookInfo.setIntroduce("暂无简介");
         }
-        String[] temp = novelUrl.split("/");
-        char c;
-        if (temp[temp.length - 1].length() == 4) {
-            c = temp[temp.length - 1].charAt(0);
-        } else {
-            c = '0';
-        }
-        bookInfo.setCoverUrl(BeQuGeService.COVER_URL + "/" + c + "/" + temp[temp.length - 1] + "/" + temp[temp.length - 1] + "s.jpg");
+        bookInfo.setCoverUrl(ZeroBookService.COVER_URL + "/" + toPinyin(bookInfo.getName(), "").toLowerCase() + ".jpg");
         bookInfo.setChapterUrl(novelUrl);
         bookInfo.setOrigin(TAG);
 
@@ -295,9 +274,9 @@ public class BiQuGeBookModelImpl extends MBaseModelImpl implements StationBookMo
     /////////////////////////////////////////////////////////////////////////////////////////////////////////////
     @Override
     public Observable<WebChapter<BookShelf>> getChapterList(BookShelf bookShelf) {
-        return getRetrofitObject(BeQuGeService.URL)
-                .create(BeQuGeService.class)
-                .getChapterList(bookShelf.getBookInfo().getChapterUrl().replace(BeQuGeService.URL, ""))
+        return getRetrofitObject(ZeroBookService.URL)
+                .create(ZeroBookService.class)
+                .getChapterList(bookShelf.getBookInfo().getChapterUrl().replace(ZeroBookService.URL, ""))
                 .flatMap((Function<String, ObservableSource<WebChapter<BookShelf>>>) s -> analyzeChapterList(s, bookShelf))
                 .subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread());
@@ -305,7 +284,7 @@ public class BiQuGeBookModelImpl extends MBaseModelImpl implements StationBookMo
 
     private Observable<WebChapter<BookShelf>> analyzeChapterList(final String s, final BookShelf bookShelf) {
         return Observable.create(e -> {
-            bookShelf.setTag(BeQuGeService.URL);
+            bookShelf.setTag(ZeroBookService.URL);
             WebChapter<List<ChapterList>> temp = analyzeChapterList(s, bookShelf.getNoteUrl());
             bookShelf.getBookInfo().setChapterlist(temp.getData());
             e.onNext(new WebChapter<>(bookShelf, temp.getNext()));
@@ -316,15 +295,16 @@ public class BiQuGeBookModelImpl extends MBaseModelImpl implements StationBookMo
     private WebChapter<List<ChapterList>> analyzeChapterList(String s, String novelUrl) {
 
         Document doc = Jsoup.parse(s);
-        Elements chapterList = doc.getElementById("section-list").getElementsByTag("li");
+        Elements chapterList = doc.getElementsByClass("section-list fix").get(1).getElementsByTag("li");
         List<ChapterList> chapters = new ArrayList<>();
         for (int i = 0; i < chapterList.size(); i++) {
             ChapterList temp = new ChapterList();
-            temp.setDurChapterUrl(BeQuGeService.URL + "/" + novelUrl.split("/")[3] + "/" + chapterList.get(i).getElementsByTag("a").attr("href"));   //id
+            temp.setDurChapterUrl(ZeroBookService.URL + chapterList.get(i).getElementsByTag("a").attr("href"));   //id
+            Log.d(TAG, "analyzeChapterList: " + temp.getDurChapterUrl());
             temp.setDurChapterIndex(i);
             temp.setDurChapterName(chapterList.get(i).getElementsByTag("a").text());
             temp.setNoteUrl(novelUrl);
-            temp.setTag(BeQuGeService.URL);
+            temp.setTag(ZeroBookService.URL);
 
             chapters.add(temp);
         }
@@ -334,9 +314,9 @@ public class BiQuGeBookModelImpl extends MBaseModelImpl implements StationBookMo
     /////////////////////////////////////////////////////////////////////////////////////////////////////////////
     @Override
     public Observable<BookContent> getBookContent(String durChapterUrl, int durChapterIndex) {
-        return getRetrofitObject(BeQuGeService.URL)
-                .create(BeQuGeService.class)
-                .getBookContent(durChapterUrl.replace(BeQuGeService.URL, ""))
+        return getRetrofitObject(ZeroBookService.URL)
+                .create(ZeroBookService.class)
+                .getBookContent(durChapterUrl.replace(ZeroBookService.URL, ""))
                 .flatMap((Function<String, ObservableSource<BookContent>>) s -> analyzeBookContent(s, durChapterUrl, durChapterIndex));
     }
 
@@ -346,7 +326,7 @@ public class BiQuGeBookModelImpl extends MBaseModelImpl implements StationBookMo
             BookContent bookContent = new BookContent();
             bookContent.setDurChapterIndex(durChapterIndex);
             bookContent.setDurChapterUrl(durChapterUrl);
-            bookContent.setTag(BeQuGeService.URL);
+            bookContent.setTag(ZeroBookService.URL);
             try {
                 Document doc = Jsoup.parse(s);
                 List<TextNode> contentEs = doc.getElementById("content").textNodes();
