@@ -2,19 +2,18 @@ package com.ebook.basebook.view;
 
 import android.animation.Animator;
 import android.animation.ObjectAnimator;
-import android.annotation.TargetApi;
+import android.annotation.SuppressLint;
 import android.content.Context;
 import android.content.res.TypedArray;
-import android.os.Build;
 import android.os.CountDownTimer;
 import android.util.AttributeSet;
 import android.view.MotionEvent;
-import android.view.View;
 import android.view.ViewGroup;
 import android.view.ViewTreeObserver;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
 
+import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
@@ -22,20 +21,19 @@ import androidx.recyclerview.widget.RecyclerView;
 import com.ebook.basebook.R;
 import com.ebook.common.util.DisplayUtil;
 
+import java.util.Objects;
+
 public class RecyclerViewBar extends LinearLayout {
     public static long SLIDE_ANIM_TIME = 800;
-
+    private final TimeCountDown timeCountDown = new TimeCountDown();
     private ImageView ivSlider;
     private int sliderHeight = DisplayUtil.dip2px(35f);
-
     private RecyclerView recyclerView;
-
     private Animator slideIn;
     private Animator slideOut;
     private float finalY = -10000;
-    private TimeCountDown timeCountDown = new TimeCountDown();
     private int height = 0;
-    private ViewTreeObserver.OnGlobalLayoutListener layoutInitListener = new ViewTreeObserver.OnGlobalLayoutListener() {
+    private final ViewTreeObserver.OnGlobalLayoutListener layoutInitListener = new ViewTreeObserver.OnGlobalLayoutListener() {
         @Override
         public void onGlobalLayout() {
             if (getHeight() > 0) {
@@ -67,7 +65,6 @@ public class RecyclerViewBar extends LinearLayout {
         init(attrs);
     }
 
-    @TargetApi(Build.VERSION_CODES.LOLLIPOP)
     public RecyclerViewBar(Context context, AttributeSet attrs, int defStyleAttr, int defStyleRes) {
         super(context, attrs, defStyleAttr, defStyleRes);
         init(attrs);
@@ -75,7 +72,7 @@ public class RecyclerViewBar extends LinearLayout {
 
     private void init(AttributeSet attrs) {
         setOrientation(VERTICAL);
-        TypedArray a = getContext().obtainStyledAttributes(attrs, R.styleable.RecyclerViewBar);
+        @SuppressLint("Recycle") TypedArray a = getContext().obtainStyledAttributes(attrs, R.styleable.RecyclerViewBar);
         sliderHeight = a.getDimensionPixelSize(R.styleable.RecyclerViewBar_slider_height, sliderHeight);
         int paddingLeft = a.getDimensionPixelSize(R.styleable.RecyclerViewBar_slider_paddingLeft, 0);
         int paddingRight = a.getDimensionPixelSize(R.styleable.RecyclerViewBar_slider_paddingRight, 0);
@@ -94,43 +91,41 @@ public class RecyclerViewBar extends LinearLayout {
         RecyclerViewBar.this.getViewTreeObserver().addOnGlobalLayoutListener(layoutInitListener);
     }
 
+    @SuppressLint("ClickableViewAccessibility")
     private void initIvSlider() {
-        ivSlider.setOnTouchListener(new OnTouchListener() {
-            @Override
-            public boolean onTouch(View v, MotionEvent event) {
-                int action = event.getAction();
-                switch (action) {
-                    case MotionEvent.ACTION_DOWN:
-                        finalY = event.getY();
-                        return true;
-                    case MotionEvent.ACTION_MOVE:
-                        if (finalY >= 0) {
-                            float tempY = event.getY();
-                            float durY = tempY - finalY;
-                            updateSlider(durY);
+        ivSlider.setOnTouchListener((v, event) -> {
+            int action = event.getAction();
+            switch (action) {
+                case MotionEvent.ACTION_DOWN:
+                    finalY = event.getY();
+                    return true;
+                case MotionEvent.ACTION_MOVE:
+                    if (finalY >= 0) {
+                        float tempY = event.getY();
+                        float durY = tempY - finalY;
+                        updateSlider(durY);
 
-                            showSlide();
-                        } else {
-                            finalY = event.getY();
-                        }
+                        showSlide();
+                    } else {
+                        finalY = event.getY();
+                    }
+                    return true;
+                case MotionEvent.ACTION_UP:
+                    if (finalY >= 0) {
+                        finalY = -10000;
+                        timeCountDown.cancel();
+                        timeCountDown.start();
                         return true;
-                    case MotionEvent.ACTION_UP:
-                        if (finalY >= 0) {
-                            finalY = -10000;
-                            timeCountDown.cancel();
-                            timeCountDown.start();
-                            return true;
-                        }
-                        break;
-                    default:
-                        if (finalY >= 0) {
-                            finalY = -10000;
-                            return true;
-                        }
-                        break;
-                }
-                return false;
+                    }
+                    break;
+                default:
+                    if (finalY >= 0) {
+                        finalY = -10000;
+                        return true;
+                    }
+                    break;
             }
+            return false;
         });
     }
 
@@ -143,8 +138,8 @@ public class RecyclerViewBar extends LinearLayout {
             finalMarginTop = getHeight() - sliderHeight;
         }
         if (recyclerView != null) {
-            int position = Math.round(finalMarginTop / (getHeight() - sliderHeight) * (recyclerView.getAdapter().getItemCount() - 1));
-            ((LinearLayoutManager) recyclerView.getLayoutManager()).scrollToPositionWithOffset(position, 0);
+            int position = Math.round(finalMarginTop / (getHeight() - sliderHeight) * (Objects.requireNonNull(recyclerView.getAdapter()).getItemCount() - 1));
+            ((LinearLayoutManager) Objects.requireNonNull(recyclerView.getLayoutManager())).scrollToPositionWithOffset(position, 0);
         }
 
         l.topMargin = Math.round(finalMarginTop);
@@ -156,7 +151,7 @@ public class RecyclerViewBar extends LinearLayout {
         if (this.recyclerView != null) {
             this.recyclerView.addOnScrollListener(new RecyclerView.OnScrollListener() {
                 @Override
-                public void onScrollStateChanged(RecyclerView recyclerView, int newState) {
+                public void onScrollStateChanged(@NonNull RecyclerView recyclerView, int newState) {
                     super.onScrollStateChanged(recyclerView, newState);
                     if (newState != 0) {
                         showSlide();
@@ -167,16 +162,16 @@ public class RecyclerViewBar extends LinearLayout {
                 }
 
                 @Override
-                public void onScrolled(RecyclerView recyclerView, int dx, int dy) {
+                public void onScrolled(@NonNull RecyclerView recyclerView, int dx, int dy) {
                     super.onScrolled(recyclerView, dx, dy);
-                    scrollToPositionWithOffset(((LinearLayoutManager) recyclerView.getLayoutManager()).findFirstVisibleItemPosition());
+                    scrollToPositionWithOffset(((LinearLayoutManager) Objects.requireNonNull(recyclerView.getLayoutManager())).findFirstVisibleItemPosition());
                 }
             });
         }
     }
 
     public void scrollToPositionWithOffset(int position) {
-        if (recyclerView != null && position < recyclerView.getAdapter().getItemCount()) {
+        if (recyclerView != null && position < Objects.requireNonNull(recyclerView.getAdapter()).getItemCount()) {
             float temp = position * 1.0f / recyclerView.getAdapter().getItemCount();
             LayoutParams l = (LayoutParams) ivSlider.getLayoutParams();
             l.topMargin = Math.round(((getHeight() - sliderHeight) * temp));
