@@ -1,21 +1,20 @@
 package com.ebook.api.service;
 
 import io.reactivex.Observable;
-import retrofit2.http.Field;
-import retrofit2.http.FormUrlEncoded;
 import retrofit2.http.GET;
 import retrofit2.http.Headers;
-import retrofit2.http.POST;
+import retrofit2.http.Query;
 import retrofit2.http.Url;
 
 /**
  * @author xrn1997
  * @date 2021/6/19
  */
-public interface BeQuGeService {
+public interface ZeroBookService {
 
-    String URL = "https://www.shuangliusc.com";
-    String COVER_URL = "https://img.shuangliusc.com";
+    String URL = "http://www.bbzayy.com";
+    String COVER_URL = "http://www.sundung.com";
+    String SEARCH_URL = "http://sou.bbzayy.com";
 
     @GET
     @Headers({"Accept:text/html,application/xhtml+xml,application/xml",
@@ -33,15 +32,14 @@ public interface BeQuGeService {
             "Cache-Control:no-cache"})
     Observable<String> getBookInfo(@Url String url);
 
-    @FormUrlEncoded
-    @POST("/modules/article/search.php")
+    @GET("/search.aspx")
     @Headers({"Accept:text/html,application/xhtml+xml,application/xml",
             "User-Agent:Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.101 Safari/537.36 Edg/91.0.864.48",
-            "Content-Type:application/x-www-form-urlencoded; charset=GBK",
+            "Content-Type:application/x-www-form-urlencoded; charset=UTF-8",
             "Accept-Charset:UTF-8",
             "Connection:close",
             "Cache-Control:no-cache"})
-    Observable<String> searchBook(@Field(value = "searchkey", encoded = true) String content);
+    Observable<String> searchBook(@Query(value = "key", encoded = true) String content, @Query("page") int page, @Query("siteid") String siteid);
 
     @GET
     @Headers({"Accept:text/html,application/xhtml+xml,application/xml",
